arquitetura refatorada
